Sync the AetherKiri engine fork with the 81 commits upstream published
since the branch point (7a1a7d9).

The two lines had rewritten the same core files, so the merge was run with
a rustfmt-normalised three-way driver: both sides are formatted with the
same style before comparison, which removed ~460 formatting-only conflicts
and left 266 real ones. Every conflict was resolved by hand; the Aether
host integration, the game compatibility fixes and the debug-build profile
overrides stay on the overlay side, and upstream's work is taken wherever
it does not overlap them.

Taken from upstream:

* na_wmv_player (VC-1 video, WMA/Pro tables, decoder fixes) and the
  shion-xfile validation work.
* siglus_cfx_decompiler HLSL/WGSL improvements and benchmark support.
* siglus_key_recovery hardening (KeyStatus, unverifiable packs) plus the
  resource/gameexe lookup fixes that came with it.
* Scene string codec/MDL detection, scene_pck reader updates and the
  key.toml string-encryption override (kept at the overlay default of
  automatic MDL detection).
* Edition 2024 and the workspace lint table, `[lints] workspace = true`,
  the release profile cleanup and the assorted engine fixes that live in
  files the overlay never touched.

Kept on the overlay side (upstream's refactors there are follow-up work):

* vm.rs, runtime/mod.rs, runtime/forms/{stage,syscom,global}.rs,
  original_save.rs, ui.rs, render/mod.rs, image_manager.rs, host.rs,
  pump_host.rs, desktop_* and bin/siglus_engine.rs, i.e. the files that
  carry the Aether host bridge, the emote/JPEG/lang-variant integration
  and the debug display path.
* winit stays at 0.30 because the overlay window stack (Renderer, host,
  message boxes) is built against the 0.30 window API; the upstream-only
  modules that need the 0.31 beta host traits (ime, desktop_config,
  desktop_icon, desktop_chihaya_bench) and the x_viewer rewrite are not
  compiled by this branch yet.

Verified: cargo check --workspace and the siglus_scene_vm test suite.

+++ crates/siglus_scene_vm/src/runtime/forms/codes.rs
@@ -3169,7 +3169,7 @@ pub mod excall_op {
     pub const FRAME_ACTION_CH: i32 = 10;
     pub const F: i32 = 7;
     pub const SCRIPT: i32 = 13;
-    pub const ARRAY_INDEX: i32 = super::ELM_ARRAY as i32;
+    pub const ARRAY_INDEX: i32 = super::ELM_ARRAY;
     pub const OP_0: i32 = 0;
     pub const OP_1: i32 = 1;
     pub const OP_2: i32 = 2;
